+++ kimchi/src/circuits/gate.rs
@@ -5,8 +5,8 @@ use crate::{
         argument::{Argument, ArgumentEnv},
         constraints::ConstraintSystem,
         polynomials::{
-            complete_add, endomul_scalar, endosclmul, foreign_field_add, foreign_field_mul,
-            poseidon, range_check, turshi, varbasemul,
+            complete_add, endomul_scalar, endosclmul, foreign_field_add, foreign_field_mul, keccak,
+            poseidon, range_check, rot, turshi, varbasemul, xor,
         },
         wires::*,
     },
@@ -21,11 +21,7 @@ use serde_with::serde_as;
 use std::io::{Result as IoResult, Write};
 use thiserror::Error;
 
-use super::{
-    argument::ArgumentWitness,
-    expr,
-    polynomials::{keccak, rot, xor},
-};
+use super::{argument::ArgumentWitness, expr};
 
 /// A row accessible from a given row, corresponds to the fact that we open all polynomials
 /// at `zeta` **and** `omega * zeta`.
@@ -329,8 +325,12 @@ impl<F: PrimeField + SquareRootField> CircuitGate<F> {
             }
             GateType::Xor16 => xor::Xor16::constraint_checks(&env, &mut cache),
             GateType::Rot64 => rot::Rot64::constraint_checks(&env, &mut cache),
-            GateType::KeccakRound => keccak::KeccakRound::constraint_checks(&env, &mut cache),
-            GateType::KeccakSponge => keccak::KeccakSponge::constraint_checks(&env, &mut cache),
+            GateType::KeccakRound => {
+                keccak::circuitgates::KeccakRound::constraint_checks(&env, &mut cache)
+            }
+            GateType::KeccakSponge => {
+                keccak::circuitgates::KeccakSponge::constraint_checks(&env, &mut cache)
+            }
         };
 
         // Check for failed constraints
